ui/ops: extract PostgresNodeStatus into a proper SFC
Replace inline render-function component with a template-based SFC,
addressing reviewer feedback. Move pg-* scoped styles to the new file.
